Data piping
Edited code to make it more comfortable to run.
Maps should be stored in the repository folder and named consistently.
If the naming is respected, by editing the first line of the script in to the name of the map, everything else should automatically adjust.
Information (geometry )is read from the SUMO_HOME folder, which is dynamically found.
This design comes from the idea of having an independent folder with runners and outputs that can be shared via dropbox or Github, upon sharing, the folder should search and find local installation.

Extra content:
Added independent runners for
-Basic working case with included improvements
-Basic Kings Cross area implementation (working albeit in a rudimentary manner)
-Basic cross implementation (network parser broken at the time of this upload)