Doublecross Update
Added new simple network based on 3 intersections in line.
Tweaked deployment files. They should work with new geometry.
Added a line to the runner to make sure global environment variables are set.
Modified some cells in the BasicQ_cross s.t. now parser reads correctly and Qlearners are deployed in the right places.

To do: Tweak reward function and actions to be a part of a semi-fixed cycle in length.