07.08.2018
Found network parser bug: If a single intersection with a single light and a single program is defined, the ID returned is a "str" instead of "int".
Workaround: define 2 programs, they will be overwritten by the external controller logic. The downside is that in this case, the number of possible phases should be defined manually.
Multiple junction maps shouldn't be affected.

Added extra maps.