Fixed Network Parser
Fixed a persistent bug in the network parser that prevented reward from being calculated in certain lines.
As a result, performance of the algorithm has sky-rocketed.
Also included some basic measure by plotting the historical reward. If the controller is working as intended, the reward should be increasing (becoming less negative) over time.